refactor(perf): split /ai-architect-academy + /investor into server shell + motion island
29th + 30th in the pattern series.
