+++ .gitignore
@@ -109,7 +109,7 @@ frankxai/
 # siblings under C:/Users/frank/, not nested. Frank to relocate when convenient.
 frankx/
 frankx-os/
-ai-architect-academy/
+/ai-architect-academy/
 education/context-engineering-for-creators/
 claude-skills/
 products/mcp-doctor/
